fixed propagation bug

+++ GraphicsSandBox/Application.cpp
@@ -149,7 +149,7 @@ void Application::DoFrame()
     {
         if (!wnd.CursorEnabled())
         {
-            camera.Rotate(delta->x, delta->y);
+            camera.Rotate((float)delta->x, (float)delta->y);
         }
     }
